AssetStore crash fixes + cleanup (#68)
* Fix iterator invalidation in AssetStore

- Ensure locks are not released until iterator derefs are complete.

* WIP: AssetStore overhaul

* Revert "Add TextureWrap, refactor Material"

This reverts commit 17f6493a05ecfc9f804e168d317a7b6a0ef44798.

* WIP: cleanup

+++ levk/window/src/impl.cpp
@@ -98,7 +98,7 @@ UniqueGlfwWin Manager::Impl::make(CreateInfo const& info) {
 		return {};
 	}
 	if (style == Style::eDecoratedWindow || style == Style::eBorderlessWindow) {
-		glm::vec2 pos{};
+		glm::ivec2 pos{};
 		if (info.config.position) {
 			pos = *info.config.position;
 		} else {
